Replaced 'long long' type with 'int64_t' exact-width integer type

+++ tests/bounds/IfElse/bound_naive/IfElse.h
@@ -1,3 +1,3 @@
 #include "Util.h"
-long long f(long long x);
+int64_t f(int64_t x);
 int main(int argc, char** args);
