+++ tests/code/CoinGame4/naive_dealloc/CoinGame4.h
@@ -1,7 +1,7 @@
 #include "Util.h"
-void play(FILE* sys, _DECL_1DARRAY_PARAM(moves), _DECL_DEALLOC_PARAM(moves), long long n);
-long long findMin(long long a, long long b);
-long long findMax(long long a, long long b);
-long long* pickCoin(_DECL_1DARRAY_PARAM(moves), _DECL_DEALLOC_PARAM(moves), long long i, long long j, long long n, long long x, long long y, long long z, _DECL_1DARRAYSIZE_PARAM_CALLBYREFERENCE);
-long long* findMoves(_DECL_1DARRAY_PARAM(moves), _DECL_DEALLOC_PARAM(moves), long long n, _DECL_1DARRAYSIZE_PARAM_CALLBYREFERENCE);
+void play(FILE* sys, _DECL_1DARRAY_PARAM(moves), _DECL_DEALLOC_PARAM(moves), int64_t n);
+int64_t findMin(int64_t a, int64_t b);
+int64_t findMax(int64_t a, int64_t b);
+int64_t* pickCoin(_DECL_1DARRAY_PARAM(moves), _DECL_DEALLOC_PARAM(moves), int64_t i, int64_t j, int64_t n, int64_t x, int64_t y, int64_t z, _DECL_1DARRAYSIZE_PARAM_CALLBYREFERENCE);
+int64_t* findMoves(_DECL_1DARRAY_PARAM(moves), _DECL_DEALLOC_PARAM(moves), int64_t n, _DECL_1DARRAYSIZE_PARAM_CALLBYREFERENCE);
 int main(int argc, char** args);
