+++ tests/code/FunctionCall2/copy_reduced_dealloc/FunctionCall2.h
@@ -1,3 +1,3 @@
 #include "Util.h"
-long long* func(_DECL_1DARRAY_PARAM(s), _DECL_DEALLOC_PARAM(s), long long n, _DECL_1DARRAYSIZE_PARAM_CALLBYREFERENCE);
+int64_t* func(_DECL_1DARRAY_PARAM(s), _DECL_DEALLOC_PARAM(s), int64_t n, _DECL_1DARRAYSIZE_PARAM_CALLBYREFERENCE);
 int main(int argc, char** args);
