+++ tests/code/gcd/naive/gcd.h
@@ -1,5 +1,5 @@
 #include "Util.h"
-typedef long long nat;
-long long gcd(long long a, long long b);
-long long gcd_cached(long long a, long long b, long long n, _DECL_1DARRAY_PARAM(gcds));
+typedef int64_t nat;
+int64_t gcd(int64_t a, int64_t b);
+int64_t gcd_cached(int64_t a, int64_t b, int64_t n, _DECL_1DARRAY_PARAM(gcds));
 int main(int argc, char** args);
